Rewrite scheduler context switch code (#1786)
* Rewrite scheduler context switch code

* Fix race in UnmapIpcRestorePermission

* Fix thread exit issue that could leave the scheduler in a invalid state

* Change context switch method to not wait on guest thread, remove spin wait, use SignalAndWait to pass control

* Remove multi-core setting (it is always on now)

* Re-enable assert

* Remove multicore from default config and schema

* Fix race in KTimeManager

+++ Ryujinx.Common/PerformanceCounter.cs
@@ -32,7 +32,7 @@ namespace Ryujinx.Common
         public static long TicksPerMillisecond { get; }
 
         /// <summary>
-        /// Gets the number of milliseconds elapsed since the system started.
+        /// Gets the number of ticks elapsed since the system started.
         /// </summary>
         public static long ElapsedTicks
         {
@@ -76,7 +76,7 @@ namespace Ryujinx.Common
             TicksPerHour        = TicksPerMinute * 60;
             TicksPerDay         = TicksPerHour * 24;
 
-            _ticksToNs = 1000000000.0 / (double)Stopwatch.Frequency;
+            _ticksToNs = 1000000000.0 / Stopwatch.Frequency;
         }
     }
 }
